agent/requirements.txt
websockets==12.0
pyyaml==6.0
requests==2.31.0
